Move all test code to one file for now
